% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_distinct_counts}
\alias{check_distinct_counts}
\title{QC-08: Report distinct value counts for character columns}
\usage{
check_distinct_counts(df, config)
}
\description{
QC-08: Report distinct value counts for character columns
}
\keyword{internal}
