% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_duplicate_rows}
\alias{check_duplicate_rows}
\title{QC-03: Check for fully-duplicate rows}
\usage{
check_duplicate_rows(df, config)
}
\description{
QC-03: Check for fully-duplicate rows
}
\keyword{internal}
