% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_inferred_types}
\alias{check_inferred_types}
\title{QC-06: Report inferred column types}
\usage{
check_inferred_types(df, config)
}
\description{
QC-06: Report inferred column types
}
\keyword{internal}
