% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_missing_rate}
\alias{check_missing_rate}
\title{QC-01: Check missing rate per column}
\usage{
check_missing_rate(df, config)
}
\description{
QC-01: Check missing rate per column
}
\keyword{internal}
