% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_numeric_stats}
\alias{check_numeric_stats}
\title{QC-07: Report numeric summary statistics}
\usage{
check_numeric_stats(df, config)
}
\description{
QC-07: Report numeric summary statistics
}
\keyword{internal}
