% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_column_order}
\alias{compare_column_order}
\title{CP-08: Check column order consistency between deliveries}
\usage{
compare_column_order(df_current, df_previous, config)
}
\description{
CP-08: Check column order consistency between deliveries
}
\keyword{internal}
