% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_dropped_values}
\alias{compare_dropped_values}
\title{CP-06: Detect dropped distinct values in character columns}
\usage{
compare_dropped_values(df_current, df_previous, config)
}
\description{
CP-06: Detect dropped distinct values in character columns
}
\keyword{internal}
