% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_new_values}
\alias{compare_new_values}
\title{CP-05: Detect new distinct values in character columns}
\usage{
compare_new_values(df_current, df_previous, config)
}
\description{
CP-05: Detect new distinct values in character columns
}
\keyword{internal}
