% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_non_numeric_rate}
\alias{compare_non_numeric_rate}
\title{CP-07: Compare non-numeric rate in numeric columns between deliveries}
\usage{
compare_non_numeric_rate(df_current, df_previous, config)
}
\description{
CP-07: Compare non-numeric rate in numeric columns between deliveries
}
\keyword{internal}
