% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_numeric_mean}
\alias{compare_numeric_mean}
\title{CP-04: Compare numeric column means between deliveries}
\usage{
compare_numeric_mean(df_current, df_previous, config)
}
\description{
CP-04: Compare numeric column means between deliveries
}
\keyword{internal}
