% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_row_count}
\alias{compare_row_count}
\title{CP-01: Compare row count between deliveries}
\usage{
compare_row_count(df_current, df_previous, config)
}
\description{
CP-01: Compare row count between deliveries
}
\keyword{internal}
