% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_schema}
\alias{compare_schema}
\title{CP-02: Detect schema differences between deliveries}
\usage{
compare_schema(df_current, df_previous, config)
}
\description{
CP-02: Detect schema differences between deliveries
}
\keyword{internal}
