% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render the HTML data quality report}
\usage{
render_report(
  dataset_name,
  file_name,
  file_path,
  df,
  qc_results,
  cp_results,
  custom_results,
  snapshot_history,
  config,
  col_stats = NULL,
  output_dir,
  open_report = TRUE
)
}
\description{
Render the HTML data quality report
}
\keyword{internal}
