% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_allowed_values}
\alias{check_allowed_values}
\title{QC-09: Check for values outside the allowed set}
\usage{
check_allowed_values(df, config)
}
\description{
QC-09: Check for values outside the allowed set
}
\keyword{internal}
