% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_empty_column}
\alias{check_empty_column}
\title{QC-02: Check for entirely empty columns}
\usage{
check_empty_column(df, config)
}
\description{
QC-02: Check for entirely empty columns
}
\keyword{internal}
