% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_key_uniqueness}
\alias{check_key_uniqueness}
\title{QC-12: Check uniqueness of key columns}
\usage{
check_key_uniqueness(df, config)
}
\description{
QC-12: Check uniqueness of key columns
}
\keyword{internal}
