% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_min_row_count}
\alias{check_min_row_count}
\title{QC-14: Check minimum row count threshold}
\usage{
check_min_row_count(df, config)
}
\description{
QC-14: Check minimum row count threshold
}
\keyword{internal}
