% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_non_numeric}
\alias{check_non_numeric}
\title{QC-11: Check non-numeric rate in numeric columns}
\usage{
check_non_numeric(df, config)
}
\description{
QC-11: Check non-numeric rate in numeric columns
}
\keyword{internal}
