% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_numeric_bounds}
\alias{check_numeric_bounds}
\title{QC-10: Check for out-of-range numeric values}
\usage{
check_numeric_bounds(df, config)
}
\description{
QC-10: Check for out-of-range numeric values
}
\keyword{internal}
