% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_pattern}
\alias{check_pattern}
\title{QC-13: Check values against a regex pattern}
\usage{
check_pattern(df, config)
}
\description{
QC-13: Check values against a regex pattern
}
\keyword{internal}
