% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_generic.R
\name{check_schema_contract}
\alias{check_schema_contract}
\title{SC-01/SC-02: Check columns against expected schema contract}
\usage{
check_schema_contract(df, config)
}
\description{
SC-01/SC-02: Check columns against expected schema contract
}
\keyword{internal}
