% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overall_status}
\alias{overall_status}
\title{Compute the worst status across a list of dq_result objects}
\usage{
overall_status(results)
}
\description{
Compute the worst status across a list of dq_result objects
}
\keyword{internal}
